DeveloperTool/AiDoc.WebNew/python/requirements.txt
# MCP server for Db2 LUW docs RAG (Cursor)
mcp[stdio]>=1.0.0
chromadb>=0.4.0
onnxruntime>=1.16.0
tokenizers>=0.15.0
pypdf>=4.0.0
python-docx>=1.1.0
